fix: #8483 payments ui tweaks
