+++ packages/payments-portal/src/components/payment/__styles__/index.ts
@@ -1,0 +1,6 @@
+import { styled } from '@linaria/react'
+
+export const MaxWidthContainer = styled.div`
+  max-width: 1440px;
+  margin: 0 auto;
+`
